pyproject.toml
[tool.poetry]
name = "oreo"
version = "1.0.0.0"
description = "The Stuffing for Other Functions!"
authors = ["sylvorg <jeet.ray@syvl.org>"]
license = "OREO"
include = [
	"oreo/**/*.py",
	"oreo/**/*.hy",
]

[tool.poetry.dependencies]
python = "^3.10"
cryptography = "*"
hy = "^0.25.0"
hyrule = "^0.2"
more-itertools = "*"
nixpkgs = "*"
pytest = "*"
valiant = { git = "https://github.com/syvlorg/valiant.git", branch = "main" }

[tool.poetry.dev-dependencies]
poetry = "*"
poetry2setup = "*"
pytest-hy = { git = "https://github.com/syvlorg/pytest-hy.git", branch = "master" }
pytest-randomly = "*"
pytest-parametrized = "*"
pytest-sugar = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = [
	"--strict-markers",
]
markers = [
	"either",
	"first_last_n",
	"flatten",
	"gensing",
	"get_un_mangled",
	"let_cwd",
	"ls",
	"multi",
	"palindrome",
	"recursive_unmangle",
	"remove_prefix_n",
	"remove_suffix_n",
	"with_cwd",
]
